Trait attohttpc::body::Body

pub trait Body {
    // Required methods
    fn kind(&mut self) -> IoResult<BodyKind>;
    fn write<W: Write>(&mut self, writer: W) -> IoResult<()>;

    // Provided method
    fn content_type(&mut self) -> IoResult<Option<String>> { ... }
}

A generic rewindable request body

Required Methods

fn kind(&mut self) -> IoResult<BodyKind>

Determine the kind of the request body

fn write<W: Write>(&mut self, writer: W) -> IoResult<()>

Write out the request body into the given writer

This method can be called multiple times if a request is redirected.

Provided Methods

fn content_type(&mut self) -> IoResult<Option<String>>

Gets the content type this body is tied to if it has one.