Struct attohttpc::RequestBuilder

pub struct RequestBuilder<B = [u8; 0]> { /* fields omitted */ }

Request is the main way of performing requests.

You can create a RequestBuilder the hard way using the new or try_new method, or use one of the simpler constructors available in the crate root, such as get post, etc.

Implementations

impl RequestBuilder

pub fn new<U>(method: Method, base_url: U) -> Self where
U: AsRef<str>,

Create a new Request with the base URL and the given method.

Panics

Panics if the base url is invalid or if the method is CONNECT.

pub fn try_new<U>(method: Method, base_url: U) -> Result<Self> where
U: AsRef<str>,

Try to create a new RequestBuilder.

If the base URL is invalid, an error is returned. If the method is CONNECT, an error is also returned. CONNECT is not yet supported.

impl<B> RequestBuilder<B>

pub fn param<K, V>(self, key: K, value: V) -> Self where
K: AsRef<str>,
V: ToString,

Associate a query string parameter to the given value.

The same key can be used multiple times.

pub fn params<P, K, V>(self, pairs: P) -> Self where
P: IntoIterator,
P::Item: Borrow<(K, V)>,
K: AsRef<str>,
V: ToString,

Associated a list of pairs to query parameters.

The same key can be used multiple times.

Example

attohttpc::get("http://foo.bar").params(&[("p1", "v1"), ("p2", "v2")]);

pub fn header<H, V>(self, header: H, value: V) -> Self where
H: IntoHeaderName,
V: TryInto<HeaderValue>,
Error: From<V::Error>,

Modify a header for this Request.

If the header is already present, the value will be replaced. If you wish to append a new header, use header_append.

Panics

This method will panic if the value is invalid.

pub fn header_append<H, V>(self, header: H, value: V) -> Self where
H: IntoHeaderName,
V: TryInto<HeaderValue>,
Error: From<V::Error>,

Modify a header for this Request.

If the header is already present, the value will be replaced. If you wish to append a new header, use header_append.

Panics

This method will panic if the value is invalid.

pub fn try_header<H, V>(self, header: H, value: V) -> Result<Self> where
H: IntoHeaderName,
V: TryInto<HeaderValue>,
Error: From<V::Error>,

Modify a header for this Request.

If the header is already present, the value will be replaced. If you wish to append a new header, use header_append.

pub fn try_header_append<H, V>(self, header: H, value: V) -> Result<Self> where
H: IntoHeaderName,
V: TryInto<HeaderValue>,
Error: From<V::Error>,

Append a new header to this Request.

The new header is always appended to the Request, even if the header already exists.

pub fn bearer_auth(self, token: impl Into<String>) -> Self

Enable HTTP bearer authentication.

pub fn text(self, body: impl AsRef<str>) -> RequestBuilder<impl AsRef<[u8]>>

Set the body of this request to be text.

If the Content-Type header is unset, it will be set to text/plain and the carset to UTF-8.

pub fn bytes(self, body: impl AsRef<[u8]>) -> RequestBuilder<impl AsRef<[u8]>>

Set the body of this request to be bytes.

If the Content-Type header is unset, it will be set to application/octet-stream.

pub fn max_redirections(self, max_redirections: u32) -> Self

Set the maximum number of redirections this Request can perform.

pub fn follow_redirects(self, follow_redirects: bool) -> Self

Sets if this Request should follow redirects, 3xx codes.

This value defaults to true.

pub fn connect_timeout(self, duration: Duration) -> Self

Sets a connect timeout for this request.

The default is 30 seconds.

pub fn read_timeout(self, duration: Duration) -> Self

Sets a read timeout for this request.

The default is 30 seconds.

impl<B: AsRef<[u8]>> RequestBuilder<B>

pub fn prepare(self) -> PreparedRequest<B>

Create a PreparedRequest from this RequestBuilder.

Panics

Will panic if an error occurs trying to prepare the request. It shouldn't happen.

pub fn try_prepare(self) -> Result<PreparedRequest<B>>

Create a PreparedRequest from this RequestBuilder.

pub fn send(self) -> Result<Response>

Send this request directly.