Struct attohttpc::Response

pub struct Response { /* private fields */ }

Response represents a response returned by a server.

Implementations

impl Response

pub fn status(&self) -> StatusCode

Get the status code of this Response.

pub fn headers(&self) -> &HeaderMap

Get the headers of this Response.

pub fn is_success(&self) -> bool

Checks if the status code of this Response was a success code.

pub fn error_for_status(self) -> Result<Self>

Returns error variant if the status code was not a success code.

pub fn split(self) -> (StatusCode, HeaderMap, ResponseReader)

Split this Response into a tuple of StatusCode, HeaderMap, ResponseReader.

This method is useful to read the status code or headers after consuming the response.

pub fn write_to<W>(self, writer: W) -> Result<u64>where
W: Write,

Write the response to any object that implements Write.

pub fn bytes(self) -> Result<Vec<u8>>

Read the response to a Vec of bytes.

pub fn text(self) -> Result<String>

Read the response to a String.

If the charsets feature is enabled, it will try to decode the response using the encoding in the headers. If there’s no encoding specified in the headers, it will fall back to the default encoding, and if that’s also not specified, it will fall back to the default of ISO-8859-1.

If the charsets feature is disabled, this method is the same as calling text_utf8.

Note that both conversions are lossy, i.e. they will not raise errors when invalid data is encountered but output replacement characters instead.

pub fn text_utf8(self) -> Result<String>

Read the response body to a String using the UTF-8 encoding.

This method ignores headers and the default encoding.

Note that is lossy, i.e. it will not raise errors when invalid data is encountered but output replacement characters instead.