Struct attohttpc::ResponseReader

pub struct ResponseReader { /* private fields */ }

The ResponseReader is used to read the body of a response.

The ResponseReader implements Read and can be used like any other stream, but the data returned by Read are untouched bytes from the socket. This means that if a string is expected back, it could be in a different encoding than the expected one. In order to properly read text, use the charsets feature and the text or text_reader methods.

In general it’s best to avoid Reading directly from this object. Instead use the helper methods, they process the data stream properly.

Implementations

impl ResponseReader

pub fn write_to<W>(self, writer: W) -> Result<u64> where
W: Write,

Write the response to any object that implements Write.

pub fn bytes(self) -> Result<Vec<u8>>

Read the response to a Vec of bytes.

pub fn text(self) -> Result<String>

Read the response to a String.

If the charsets feature is enabled, it will try to decode the response using the encoding in the headers. If there’s no encoding specified in the headers, it will fall back to the default encoding, and if that’s also not specified, it will fall back to the default of ISO-8859-1.

If the charsets feature is disabled, this method is the same as calling text_utf8.

Note that both conversions are lossy, i.e. they will not raise errors when invalid data is encountered but output replacement characters instead.

pub fn text_utf8(self) -> Result<String>

Read the response body to a String using the UTF-8 encoding.

This method ignores headers and the default encoding.

Note that this is lossy, i.e. it will not raise errors when invalid data is encountered but output replacement characters instead.